// src/types/dto/tax.ts
export interface TaxRateResponse {
	id: string;
	name: string;
	description: string;
	code: string;
	status: ENTITY_STATUS;
	tax_rate_status: TAX_RATE_STATUS;
	tax_rate_type: TAX_RATE_TYPE;
	scope: TAX_RATE_SCOPE;
	percentage_value?: number;
	fixed_value?: number;
	metadata?: Record<string, string>;
	environment_id: string;
	tenant_id: string;
	created_at: string;
	updated_at: string;
	created_by: string;
	updated_by: string;
}